move python clients to dedicated modules
