+++ benchmarks/tangobenchmarks/client/python/command.py
@@ -1,0 +1,57 @@
+import tango
+import time
+from multiprocessing import Process, Queue
+
+import tangobenchmarks.utils as utils
+
+
+class Worker(Process):
+    """ worker instance
+    """
+
+    def __init__(self, wid, device, command, period, qresult):
+        """ constructor
+
+        :param wid: worker id
+        :type wid: :obj:`int`
+        :param device: device name
+        :type device: :obj:`str`
+        :param command: command name
+        :type command: :obj:`str`
+        :param period: time period
+        :type period: :obj:`float`
+        :param qresult: queue with result
+        :type qresult: :class:`Queue.Queue` or `queue.queue`
+
+        """
+        Process.__init__(self)
+
+        # : (:obj:`int`) worker id
+        self.__wid = wid
+        # : (:obj:`float`) time period in seconds
+        self.__period = float(period)
+        #: (:obj:`str`) device proxy
+        self.__device = device
+        #: (:obj:`str`) device command name
+        self.__command = command
+        # : (:class:`tango.DeviceProxy`) device proxy
+        self.__proxy = None
+        # : (:class:`Queue.Queue`) result queue
+        self.__qresult = qresult
+        # : (:obj:`int`) counter
+        self.__counter = 0
+
+    def run(self):
+        """ worker thread
+        """
+        if hasattr(tango.ApiUtil, 'cleanup'):
+            tango.ApiUtil.cleanup()
+        self.__proxy = tango.DeviceProxy(self.__device)
+        stime = time.time()
+        etime = stime
+        while etime - stime < self.__period:
+            self.__proxy.command_inout(self.__command)
+            etime = time.time()
+            self.__counter += 1
+        self.__qresult.put(
+            utils.Result(self.__wid, self.__counter, etime - stime))
