+++ benchmarks/tangobenchmarks/client/python/pipe_write.py
@@ -1,0 +1,68 @@
+import tango
+import time
+from multiprocessing import Process, Queue
+
+import tangobenchmarks.utils as utils
+
+
+class Worker(Process):
+    """ write worker instance
+    """
+
+    def __init__(self, wid, device, pipe, period, value, qresult):
+        """ constructor
+
+        :param wid: worker id
+        :type wid: :obj:`int`
+        :param device: device name
+        :type device: :obj:`str`
+        :param pipe: pipe name
+        :type pipe: :obj:`str`
+        :param period: time period
+        :type period: :obj:`float`
+        :param value:  pipe value
+        :type value: :class:`numpy.ndarray`
+        :param qresult: queue with result
+        :type qresult: :class:`Queue.Queue` or `queue.queue`
+
+        """
+        Process.__init__(self)
+
+        # : (:obj:`int`) worker id
+        self.__wid = wid
+        # : (:obj:`float`) time period in seconds
+        self.__period = float(period)
+        #: (:obj:`str`) device proxy
+        self.__device = device
+        #: (:obj:`str`) device pipe name
+        self.__pipe = pipe
+        #: (:obj:`float` or :class:`np.array`) device pipe value
+        self.__value = value
+        # : (:class:`tango.PipeProxy`) pipe proxy
+        self.__proxy = None
+        # : (:class:`Queue.Queue`) result queue
+        self.__qresult = qresult
+        # : (:obj:`int`) counter
+        self.__counter = 0
+        # : (:obj:`int`) error counter
+        self.__errors = 0
+
+    def run(self):
+        """ worker thread
+        """
+        if hasattr(tango.ApiUtil, 'cleanup'):
+            tango.ApiUtil.cleanup()
+        self.__proxy = tango.DeviceProxy(self.__device)
+        stime = time.time()
+        etime = stime
+        while etime - stime < self.__period:
+            try:
+                self.__proxy.write_pipe(self.__pipe, self.__value)
+            except Exception:
+                self.__errors += 1
+            else:
+                self.__counter += 1
+            etime = time.time()
+        self.__qresult.put(
+            utils.Result(self.__wid, self.__counter, etime - stime,
+                         self.__errors))
