+++ benchmarks/tangobenchmarks/writebenchmark.py
@@ -30,69 +30,7 @@ from multiprocessing import Process, Queue
 from . import release
 from . import utils
 
-
-class Worker(Process):
-    """ worker instance
-    """
-
-    def __init__(self, wid, device, attribute, period, value, qresult):
-        """ constructor
-
-        :param wid: worker id
-        :type wid: :obj:`int`
-        :param device: device name
-        :type device: :obj:`str`
-        :param attribute: attribute name
-        :type attribute: :obj:`str`
-        :param period: time period
-        :type period: :obj:`float`
-        :param value:  attribute value
-        :type value: :class:`numpy.ndarray`
-        :param qresult: queue with result
-        :type qresult: :class:`Queue.Queue` or `queue.queue`
-
-        """
-        Process.__init__(self)
-
-        # : (:obj:`int`) worker id
-        self.__wid = wid
-        # : (:obj:`float`) time period in seconds
-        self.__period = float(period)
-        #: (:obj:`str`) device proxy
-        self.__device = device
-        #: (:obj:`str`) device attribute name
-        self.__attribute = attribute
-        #: (:obj:`float` or :class:`np.array`) device attribute value
-        self.__value = value
-        # : (:class:`tango.AttributeProxy`) attribute proxy
-        self.__proxy = None
-        # : (:class:`Queue.Queue`) result queue
-        self.__qresult = qresult
-        # : (:obj:`int`) counter
-        self.__counter = 0
-        # : (:obj:`int`) error counter
-        self.__errors = 0
-
-    def run(self):
-        """ worker thread
-        """
-        if hasattr(tango.ApiUtil, 'cleanup'):
-            tango.ApiUtil.cleanup()
-        self.__proxy = tango.DeviceProxy(self.__device)
-        stime = time.time()
-        etime = stime
-        while etime - stime < self.__period:
-            try:
-                self.__proxy.write_attribute(
-                    self.__attribute, self.__value)
-            except Exception:
-                self.__errors += 1
-            else:
-                self.__counter += 1
-            etime = time.time()
-        self.__qresult.put(
-            utils.Result(self.__wid, self.__counter, etime - stime,
-                         self.__errors))
+from tangobenchmarks.client.python.write import Worker
 
 
 class WriteBenchmark(utils.Benchmark):
